handling error responses

+++ src/todos/todo.controller.ts
@@ -1,9 +1,10 @@
-import { BadRequestException, Body, ConsoleLogger, Controller, Get, HttpException, HttpStatus, Param, Patch, Post, Req, Request, Response, UseGuards, ValidationPipe } from '@nestjs/common';
+import { BadRequestException, Body, ConsoleLogger, Controller, Get, HttpException, HttpStatus, NotFoundException, Param, Patch, Post, Req, Request, Response, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from 'src/auth/guard/jwt-auth.guard';
 import { TodoService } from "./todo.service";
 import { ToDo as TodoModel, Prisma, Status } from '@prisma/client';
 import { PrismaService } from 'prisma/prisma.service';
 import { response } from 'express';
+import { NOTFOUND } from 'dns';
 
 @UseGuards(JwtAuthGuard)
 @Controller("todos")
@@ -88,7 +89,7 @@ export class TodoController {
     }
 
     @Get('/filter')
-    async filterTodo(@Request() req, @Body(ValidationPipe) searchDto: { completion: Status }): Promise<any> {
+    async filterTodo(@Request() req, @Body() searchDto: { completion: Status }): Promise<any> {
         const { userId } = req.user;
         const { completion } = searchDto;
         if (completion && !(completion in Status)) {
@@ -114,6 +115,10 @@ export class TodoController {
         const { id, completion } = updateTodoDto
         if (completion && !(completion in Status)) {
             throw new HttpException("invalid completion value", HttpStatus.BAD_REQUEST)
+        }
+
+        if (!this.todoService.findTodo({ where: { id } })) {
+            throw new NotFoundException(`Todo item with ID ${id} not found`)
         }
 
         const res = await this.todoService.updateCompletion({ id }, completion).then(response => {
@@ -142,6 +147,11 @@ export class TodoController {
         else if (title && title.trim().length === 0) {
             throw new BadRequestException("'Title' cannot be empty")
         }
+
+        if (!this.todoService.findTodo({ where: { id } })) {
+            throw new NotFoundException(`Todo item with ID ${id} not found`)
+        }
+
         const res = await this.todoService.updateContent({ id }, title ? title.trim() : title, desc).then(response => {
             // console.log(response)
             return response
@@ -162,6 +172,11 @@ export class TodoController {
     async removeTodo(@Request() req, @Body() removeTodoDto: { id: string }): Promise<any> {
         const { userId } = req.user
         const { id } = removeTodoDto
+
+        if (!this.todoService.findTodo({ where: { id } })) {
+            throw new NotFoundException(`Todo item with ID ${id} not found`)
+        }
+
         const res = await this.todoService.removeTodo({ id }, { userId }).then(response => {
             // console.log(response)
             return response
